rename tables and column template file, save $tables_and_columns_names
